fhandle: create do_filp_handle_open() helper
do_handle_open() takes care of both opening a file via its file handle,
and associating it with a file descriptor.

For io_uring, it is useful to do just the opening part separately,
because io_uring might not want to install it into the main open files
table when using fixed descriptors.

This creates a helper which will enable io_uring to do that.

Signed-off-by: Thomas Bertschinger <tahbertschinger@gmail.com>